app/phy/cpmGasprices/GoogleCloudFunctions/src/requirements.txt
pandas==1.3.5
fsspec==2021.11.1
gcsfs==2021.11.1
google-cloud-storage==1.43.0
matplotlib==3.5.1